subclass sources, move subclasses into same module as parent class
